+++ src/main/java/com/tech/bee/postservice/service/TagService.java
@@ -29,6 +29,8 @@ public class TagService {
     private final TagMapper tagMapper;
     private final TagValidator tagValidator;
 
+    private final SecurityService securityService;
+
     public TagDTO findTagByName(final String tagName){
         TagEntity tagEntity = tagRepository.findByNameIgnoreCase(tagName)
                 .orElseThrow( () -> BaseCustomException.builder().
@@ -54,6 +56,9 @@ public class TagService {
         TagEntity existingTag =  tagRepository.findByIdentifier(tagIdentifier).orElseThrow(() -> BaseCustomException.builder().
                 errors(Collections.singletonList(AppUtil.buildResourceNotFoundError(ApiConstants.KeyConstants.KEY_TAG))).httpStatus(HttpStatus.NOT_FOUND)
                 .build());
+        List<ErrorDTO> ownershipErrors = securityService.validateOwnership(existingTag.getCreatedBy());
+        if(CollectionUtils.isNotEmpty(ownershipErrors))
+            throw BaseCustomException.builder().errors(ownershipErrors).httpStatus(HttpStatus.FORBIDDEN).build();
         tagRepository.delete(existingTag);
     }
 
@@ -62,6 +67,9 @@ public class TagService {
         TagEntity existingTag =  tagRepository.findByIdentifier(tagIdentifier).orElseThrow(() -> BaseCustomException.builder().
                 errors(Collections.singletonList(AppUtil.buildResourceNotFoundError(ApiConstants.KeyConstants.KEY_TAG))).httpStatus(HttpStatus.NOT_FOUND)
                 .build());
+        List<ErrorDTO> ownershipErrors = securityService.validateOwnership(existingTag.getCreatedBy());
+        if(CollectionUtils.isNotEmpty(ownershipErrors))
+            throw BaseCustomException.builder().errors(ownershipErrors).httpStatus(HttpStatus.FORBIDDEN).build();
         AppUtil.mergeObjectsWithProperties(tagDTO , existingTag , Arrays.asList("name"));
         log.info("Tag successfully updated : new tag => {}",tagDTO);
     }
